MONGOLIAN-OPEN-SPEECH/requirements.txt
pyyaml>=5.1
hyperpyyaml
tensorboardX
torchaudio
numpy
Pillow
typeguard
tqdm